fix(universes): cast SUM(...) to ::int — UniverseSummary.IncludedOverrides bigint→int mismatch
User saw "Error: 500" on /universes. /api/universes/ returns 500
because the SQL aggregate `SUM(CASE WHEN action = 'INCLUDE' THEN 1
ELSE 0 END)` is BIGINT in Postgres, but the C# UniverseSummary record
declares IncludedOverrides / ExcludedOverrides as int. Dapper fails
the bigint→int conversion.

Fix: cast both sums to ::int in the subquery + cast the COALESCE
results too so Dapper sees an int4 column. Counts will never exceed
2^31 (override list is per universe; trader-curated).

+++ backend/TradePro.Api/Endpoints/UniverseEndpoints.cs
@@ -37,13 +37,13 @@ public static class UniverseEndpoints
                        u.symbol_count     AS SymbolCount,
                        u.fetched_at_utc   AS FetchedAtUtc,
                        u.source           AS Source,
-                       COALESCE(o.included_count, 0) AS IncludedOverrides,
-                       COALESCE(o.excluded_count, 0) AS ExcludedOverrides
+                       COALESCE(o.included_count, 0)::int AS IncludedOverrides,
+                       COALESCE(o.excluded_count, 0)::int AS ExcludedOverrides
                 FROM universes u
                 LEFT JOIN (
                     SELECT universe_name,
-                           SUM(CASE WHEN action = 'INCLUDE' THEN 1 ELSE 0 END) AS included_count,
-                           SUM(CASE WHEN action = 'EXCLUDE' THEN 1 ELSE 0 END) AS excluded_count
+                           SUM(CASE WHEN action = 'INCLUDE' THEN 1 ELSE 0 END)::int AS included_count,
+                           SUM(CASE WHEN action = 'EXCLUDE' THEN 1 ELSE 0 END)::int AS excluded_count
                     FROM universe_overrides
                     GROUP BY universe_name
                 ) o ON o.universe_name = u.name
